+++ src/index.js
@@ -1,6 +1,0 @@
-const { start } = require('./viewer');
-
-module.exports = {
-  start,
-  BundleAnalyzerPlugin: require('./BundleAnalyzerPlugin')
-};